CA-213968: Move get_nfs_timeout to nfs.py
