+++ drivers/NFSSR.py
@@ -137,7 +137,7 @@ class NFSSR(FileSR.FileSR):
         if not self._checkmount():
             self.validate_remotepath(False)
             util._testHost(self.dconf['server'], NFSPORT, 'NFSTarget')
-            io_timeout = util.get_nfs_timeout(self.session, sr_uuid)
+            io_timeout = nfs.get_nfs_timeout(self.session, sr_uuid)
             self.mount_remotepath(sr_uuid, io_timeout)
         self.attached = True
