+++ drivers/blktap2.py
@@ -39,6 +39,7 @@ import scsiutil
 from syslog import openlog, syslog
 from stat import * # S_ISBLK(), ...
 from SR import SROSError
+import nfs
 
 import resetvdis
 import vhdutil
@@ -1552,7 +1553,7 @@ class VDI(object):
             # Tapdisk needs the allocation_quantum to be in MB
             options["allocation_quantum"] = int(aq / (1024*1024))
 
-        timeout = util.get_nfs_timeout(self.target.vdi.session, sr_uuid)
+        timeout = nfs.get_nfs_timeout(self.target.vdi.session, sr_uuid)
         if timeout:
             options["timeout"] = timeout + self.TAPDISK_TIMEOUT_MARGIN
         for i in range(self.ATTACH_DETACH_RETRY_SECS):
