+++ drivers/nfs.py
@@ -21,6 +21,7 @@ import util
 import errno
 import os
 import xml.dom.minidom
+import XenAPI
 
 # The algorithm for tcp and udp (at least in the linux kernel) for
 # NFS timeout on softmounts is as follows:
@@ -220,3 +221,29 @@ def get_supported_nfs_versions(server):
         return list(cv & valid_versions)
     except:
         util.SMlog("Unable to obtain list of valid nfs versions")
+
+def get_nfs_timeout(session, sr_uuid):
+    if not isinstance(session, XenAPI.Session):
+        util.SMlog("No XAPI session for getting nfs timeout config")
+        return 0
+
+    try:
+        sr_ref = session.xenapi.SR.get_by_uuid(sr_uuid)
+        other_config = session.xenapi.SR.get_other_config(sr_ref)
+        str_val = other_config.get("nfs-timeout")
+    except XenAPI.Failure:
+        util.SMlog("Failed to get SR.other-config:nfs-timeout, ignoring")
+        return 0
+
+    if not str_val:
+        return 0
+
+    try:
+        nfs_timeout = int(str_val)
+        if nfs_timeout < 1:
+            raise ValueError
+    except ValueError:
+        util.SMlog("Invalid nfs-timeout value: %s" % str_val)
+        return 0
+
+    return nfs_timeout
